[6.4.1] refactors context serialization methods.
