+++ src/job_packager.js
@@ -1,13 +1,12 @@
 var errors = require('./errors');
 var utils = require('./utils');
 var Partitioner = require('./typed_array_partitioner');
+var contextSerializer = require('./context');
 
 var operation_names = require('./operation_names');
 operation_names = Object.keys(operation_names).map(function (k) {
   return operation_names[k];
 });
-
-var FUNCTION_REGEX = /^function[^(]*\(([^)]*)\)[^{]*\{([\s\S]*)\}$/;
 
 var JobPackager = module.exports = function (parts, elements) {
   if (!parts) {
@@ -36,7 +35,7 @@ JobPackager.prototype.generatePackages = function (operations, context) {
 
     var packageCodeArgs;
     var packageCode;
-    parseFunction(op.code, function (args, body) {
+    utils.parseFunction(op.code, function (args, body) {
       packageCodeArgs = args;
       packageCode = body;
     });
@@ -53,8 +52,7 @@ JobPackager.prototype.generatePackages = function (operations, context) {
   var partitioner = new Partitioner(this.parts);
   var partitionedElements = partitioner.partition(this.elements);
 
-  var ctx = sanitizeContext(context);
-  var strfyCtx = JSON.stringify(ctx);
+  var strfyCtx = contextSerializer.serialize(context);
   return partitionedElements.map(function (partitionedElement, index) {
     return {
       index: index,
@@ -65,42 +63,3 @@ JobPackager.prototype.generatePackages = function (operations, context) {
     };
   });
 };
-
-function parseFunction (code, callback) {  
-  var functionString = code.toString();
-  var match = functionString.match(FUNCTION_REGEX);
-  var args = match[1].split(',').map(function (p) { return p.trim(); });
-  var body = match[2];
-  callback(args, body);
-}
-
-function sanitizeContext (context) {
-  var ctx;
-  if (context) {
-    ctx = {};
-    for (var name in context) {
-      if (context.hasOwnProperty(name)) {
-        ctx[name] = sanitizeContextValue(context[name]);
-      }
-    }
-  }
-  return ctx;
-}
-
-function sanitizeContextValue (value) {
-  if (utils.isFunction(value)) {
-    var packageCodeArgs;
-    var packageCode;
-    parseFunction(value, function (args, body) {
-      packageCodeArgs = args;
-      packageCode = body;
-    });
-    return {
-      isFunction: true,
-      args: packageCodeArgs,
-      code: packageCode
-    };
-  } else { //TODO: (mati) ver que pasa con otros tipos
-    return value;
-  }
-}
