+++ src/worker_core.js
@@ -105,6 +105,28 @@ module.exports = function(event){
 };
 
 function createFunction(args, code) {
+  switch (args.length) {
+    case 0: 
+      /*jslint evil: true */
+      return new Function(code);
+    case 1: 
+      /*jslint evil: true */
+      return new Function(args[0], code);
+    case 2:
+      /*jslint evil: true */
+    return new Function(args[0], args[1], code);
+    case 3:
+      /*jslint evil: true */
+      return new Function(args[0], args[1], args[2], code);
+    case 4:
+      /*jslint evil: true */
+      return new Function(args[0], args[1], args[2], args[3], code);
+    default:
+      return createDynamicArgumentsFunction(args, code);
+  }
+}
+
+function createDynamicArgumentsFunction(args, code) {
   var fArgs = new Array(args);
   fArgs.push(code);
   return Function.prototype.constructor.apply(null, fArgs);
@@ -125,7 +147,7 @@ function createContext (context) {
 }
 
 function createContextValue (value) {
-  if (value && value.isFunction && value.args && value.code) {
+  if (value && value.__isFunction && value.args && value.code) {
     return createFunction(value.args, value.code);
   } else {
     return value;
